upgrade upgrade_go_sdk script

+++ upgrade_golang.sh
@@ -1,29 +1,54 @@
 #!/bin/bash
 
-# Remove existing go
-echo "Remove existing go"
-sudo rm -rf /usr/local/go
+# Get the latest version of Go available
+latest_version=$(curl -sL https://golang.org/VERSION?m=text)
 
-# Download latest go version
-echo "Download latest go version"
-latest_version=$(curl -L "https://golang.org/VERSION?m=text")
-curl -O "https://dl.google.com/go/$latest_version.linux-amd64.tar.gz"
+echo "latest version is $latest_version"
 
-# Extract the package
-echo "Extract the package"
-sudo tar -C /tmp -xzf $latest_version.linux-amd64.tar.gz
-sudo mv /tmp/go /usr/local
+upgrade() {
+    # Download latest go version
+    echo "Download latest go version"
+    curl -O "https://dl.google.com/go/$latest_version.linux-amd64.tar.gz"
 
-# Set environment variables
-# echo 'export GOPATH=$HOME/go' >> ~/.bashrc
-# echo 'export PATH=$PATH:/usr/local/go/bin:$GOPATH/bin' >> ~/.bashrc
-# 
-# source ~/.bashrc
+    # Extract the package
+    echo "Extract the package"
+    sudo tar -C /tmp -xzf $latest_version.linux-amd64.tar.gz
+    sudo mv /tmp/go /usr/local
 
-# Verify the installation
-echo "Verify the installation"
-go version
+    # Set environment variables
+    # echo 'export GOPATH=$HOME/go' >> ~/.bashrc
+    # echo 'export PATH=$PATH:/usr/local/go/bin:$GOPATH/bin' >> ~/.bashrc
+    #
+    # source ~/.bashrc
 
-# Remove downloaded file
-echo "Remove downloaded file"
-rm $latest_version.linux-amd64.tar.gz
+    # Verify the installation
+    echo "Verify the installation"
+    go version
+
+    # Remove downloaded file
+    echo "Remove downloaded file"
+    rm $latest_version.linux-amd64.tar.gz
+}
+
+# Check if Go is installed on the system
+if ! command -v go &> /dev/null; then
+    echo "Go is not installed on your system. Install it first."
+    upgrade
+    exit 1
+fi
+
+# Get the current version of Go installed on the system
+current_version=$(go version | awk '{print $3}')
+
+echo "current version is $current_version"
+
+# Check if the current version is the same as the latest version
+if [ "$current_version" == "$latest_version" ]; then
+    echo "Go is already up to date."
+    exit 0
+else
+    # Remove existing go
+    echo "Remove existing go"
+    sudo rm -rf /usr/local/go
+    upgrade
+fi
